Snapshot of develop 2019-01-28 (Internal repo commit 088bff2)
* 088bff2 (Fri, 25 Jan 2019 10:24:11 -0500)
| AV-4727 Moved duplicate files to links to their xilinx13_3 couterparts
* be23c3d (Wed, 23 Jan 2019 15:57:37 -0500)
| updates to sequence of enums output in ocpidev show component in table mode, also updated an error message to be more informative.  AV-5034
* 0b5786e (Wed, 23 Jan 2019 09:07:41 -0500)
| Final for the weekly. AV-3824. NoJenkins
* 0e3d180 (Wed, 23 Jan 2019 09:06:37 -0500)
| Quick test for AV-3824. NoJenkins
* 132461d (Wed, 23 Jan 2019 09:04:54 -0500)
| Quick test for AV-3824. NoJenkins
* c3c321d (Wed, 23 Jan 2019 08:58:41 -0500)
| Quick test for AV-3824. NoJenkins
* eab3a1f (Wed, 23 Jan 2019 08:45:09 -0500)
| Quick test for AV-3824. NoJenkins
* 99c573a (Wed, 23 Jan 2019 08:44:39 -0500)
| Quick test for AV-3824. NoJenkins
* 5c93dfc (Wed, 23 Jan 2019 08:43:04 -0500)
| Quick test for AV-3824. NoJenkins
* 62fb319 (Wed, 23 Jan 2019 08:40:21 -0500)
| Quick test for AV-3824. NoJenkins
* 84000ed (Wed, 23 Jan 2019 08:22:33 -0500)
| Testing. NoJenkins
* dac3899 (Wed, 23 Jan 2019 07:33:48 -0500)
| Add timestamps. NoJenkins
* 2a1e817 (Tue, 22 Jan 2019 16:46:49 -0500)
| fixed grammatical error in comment
* 2018767 (Tue, 22 Jan 2019 14:01:19 -0500)
| implemented ocpidev show component AV-5034
* de5d38c (Tue, 22 Jan 2019 13:12:14 -0500)
| AV-5090 Add build configuration and utilization reports
* 7f4b7ee (Tue, 22 Jan 2019 12:27:30 -0500)
| AV-5090 Update datasheets for timeiq adapters to new snippet location
* 7282121 (Tue, 22 Jan 2019 09:40:48 -0500)
| AV-5090 Move timeiq adapters from picoflexor project to assets
* 7e94af2 (Tue, 22 Jan 2019 07:12:22 -0500)
| Fix some compiler warnings
* 6f75b7d (Fri, 18 Jan 2019 11:55:04 -0500)
  Promoted important step to have its own square

+++ runtime/util/misc/src/CmdOption.cxx
@@ -114,7 +114,7 @@ namespace OCPI {
 	    goto cont2;
 	  }
 	  const char *eq = strchr(arg, '=');
-	  size_t len = eq ? eq - arg : strlen(arg);
+	  size_t len = eq ? static_cast<size_t>(eq - arg) : strlen(arg);
 	  Member *m = m_options;
 	  for (unsigned n = 0; n < m_nOptions; n++, m++)
 	    if (len == m_names[n].size() && !strncasecmp(arg, m_names[n].c_str(), len)) {
@@ -137,7 +137,7 @@ namespace OCPI {
       }
       m_argv = ap;
       while (*ap) ap++;
-      m_argvCount = ap - m_argv;
+      m_argvCount = static_cast<size_t>(ap - m_argv); // always positive
       if (debug) {
 	Member *m = m_options;
 	for (unsigned n = 0; n < m_nOptions; n++, m++)
@@ -177,7 +177,7 @@ namespace OCPI {
 	size_t nlwidth = 1000;
 	const char *nl = strchr(m->m_description.c_str(), '\n');
 	if (nl)
-	  nlwidth = nl - m->m_description.c_str();
+	  nlwidth = static_cast<size_t>(nl - m->m_description.c_str());
 	fprintf(stderr, "  %-*s   %-6s  %-8s   %s   %-.*s.", (int)width, m_names[n].c_str(),
 		m->m_abbrev.size() ? m->m_abbrev.c_str() : "<none>",
 		baseTypeNames[m->m_baseType],
@@ -185,7 +185,7 @@ namespace OCPI {
 	while (nl) {
 	  const char *start = ++nl;
 	  nl = strchr(start, '\n');
-	  nlwidth = nl ? nl - start : 1000;
+	  nlwidth = nl ? static_cast<size_t>(nl - start) : 1000;
 	  fprintf(stderr, "\n%*s%-.*s", (int)width + 30, "", (int)nlwidth, start);
 	}
 	if (m_defaults[n])
